moved class Segments to element/geom.py
